Initial commit for Node
So right now the node is used to represent a block, and we'll use B and
B for the actual algorithm.

The node object has some interesting stuff, like you can trace the
parent of each node, and the number of left and right nodes.

However, we still have to figure out how to calculate the lower bound
efficiently.

Right now we have the adjacency matrix that will help us a lot, but we
still have to go up nodes to figure out which nodes a leaf is
intersecting. It's not very efficient. Let's see what I can come up
with! :)